+++ assign6/src/fipaan/com/printf/FormattedObjType.java
@@ -4,7 +4,7 @@ import fipaan.com.errors.*;
 
 public enum FormattedObjType {
     Integer,
-    Double,
+    Float,
     String;
 
     public void expect(FormattedObjType val) {
@@ -12,11 +12,10 @@ public enum FormattedObjType {
              throw FError.New("Expected %s, got %s", this.toString(), val.toString());
         }
     }
-    @Override
-    public String toString() {
+    @Override public String toString() {
         switch (this) {
             case FormattedObjType.Integer: return "Integer";
-            case FormattedObjType.Double:  return "Double";
+            case FormattedObjType.Float:   return "Float";
             case FormattedObjType.String:  return "String";
         }
         throw FError.UNREACHABLE("unknown type");
